requirements.txt
flask
azure.storage.blob
gradio
pymssql
python-dotenv
